Implement Update Sequence Array parsing and record fixups.
This is where things get dirty. As NTFS requires us to fix up records, we can't continue our previous design of avoiding dynamic allocations and reading everything on demand via `io::Read`.
Instead, we now read an entire record (usually not larger than 4 KiB), fix it up, and store it in a `Vec`.
This required changes almost everywhere.
It should be noted that many non-resident attributes are not part of a record, which is why structured values are now implemented differently depending on the attribute type.

On the plus side, many structures need less borrowing now, which makes them more comfortable to use.
I have also added missing sanity checks with precise errors where appropriate.

+++ src/guid.rs
@@ -5,7 +5,7 @@ use binread::BinRead;
 use core::fmt;
 
 /// Size of a single GUID on disk (= size of all GUID fields).
-pub(crate) const GUID_SIZE: u64 = 16;
+pub(crate) const GUID_SIZE: usize = 16;
 
 #[derive(BinRead, Clone, Debug, Eq, PartialEq)]
 pub struct NtfsGuid {
